% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf.R
\name{workingdays}
\alias{workingdays}
\title{gives total number of working days}
\usage{
workingdays(StartDate, EndDate)
}
\arguments{
\item{StartDate}{is the staring date}

\item{EndDate}{is the ending date}
}
\description{
gives total number of working days
}
